requirements.txt
pycryptodome==3.12.0
